+++ src/controllers/postController.js
@@ -1,5 +1,6 @@
 import * as postService from '../services/postService.js';
 import CustomError from '../utils/CustomError.js';
+import { isValidObjectId } from '../utils/validations.js';
 
 export const createPost = async (req, res) => {
   try {
@@ -23,6 +24,10 @@ export const updatePost = async (req, res) => {
     const postId = req.params.id;
     const user = req.user;
 
+    if (!isValidObjectId(postId)) {
+      throw new CustomError('Id inválido', 400);
+    }
+
     const updatedPost = await postService.updatePost(postId, data, user);
 
     return res.status(200).json({ message: 'Postagem atualizada com sucesso!' });
@@ -38,6 +43,10 @@ export const deletePost = async (req, res) => {
   try {
     const postId = req.params.id;
     const user = req.user;
+
+    if (!isValidObjectId(postId)) {
+      throw new CustomError('Id inválido', 400);
+    }
 
     const updatedPost = await postService.deletePost(postId, user);
 
@@ -56,6 +65,10 @@ export const createCommentPost = async (req, res) => {
     const user = req.user;
     const commentData = req.body;
 
+    if (!isValidObjectId(postId)) {
+      throw new CustomError('Id inválido', 400);
+    }
+
     const updatedPost = await postService.createCommentPost(postId, commentData, user);
 
     return res.status(200).json({ message: 'Comentário adicionado com sucesso!' });
@@ -73,6 +86,10 @@ export const updateCommentPost = async (req, res) => {
     const postCommentId = req.params.id;
     const user = req.user;
 
+    if (!isValidObjectId(postCommentId)) {
+      throw new CustomError('Id inválido', 400);
+    }
+
     const updatedPostComment = await postService.updateCommentPost(postCommentId, data, user);
 
     return res.status(200).json({ message: 'Comentário atualizado com sucesso!' });
@@ -89,6 +106,10 @@ export const deleteCommentPost = async (req, res) => {
     const postCommentId = req.params.id;
     const user = req.user;
 
+    if (!isValidObjectId(postCommentId)) {
+      throw new CustomError('Id inválido', 400);
+    }
+
     const updatedPostComment = await postService.deleteCommentPost(postCommentId, user);
 
     return res.status(200).json({ message: 'Comentário deletado com sucesso!' });
@@ -104,6 +125,10 @@ export const createLikePost = async (req, res) => {
   try {
     const postId = req.params.id;
     const user = req.user;
+
+    if (!isValidObjectId(postId)) {
+      throw new CustomError('Id inválido', 400);
+    }
 
     const updatedPost = await postService.createLikePost(postId, user);
